refactor: Cha 및 Po 클래스를 LinearPiece로 추상화하여 중복 제거
